pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-hotel_receptionist"
version = "0.1.0"
description = "Hotel Receptionist environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server, HTTP client, Gradio web UI)
    "openenv-core[core]>=0.2.2,<0.3.0",
    # HTTP client for inference script
    "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m hotel_receptionist.server.app
server = "hotel_receptionist.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["hotel_receptionist", "hotel_receptionist.server"]
package-dir = { "hotel_receptionist" = ".", "hotel_receptionist.server" = "server" }